+++ table/table_builder.cc
@@ -136,7 +136,7 @@ void TableBuilder::Flush() {
   }
 }
 
-static bool GoodCompressionRatio(int compressed_size, int raw_size) {
+static bool GoodCompressionRatio(size_t compressed_size, size_t raw_size) {
   // Check to see if compressed less than 12.5%
   return compressed_size < raw_size - (raw_size / 8u);
 }
